requirements.txt
torch
torchvision
timm
numpy
scikit-learn
pillow
matplotlib
streamlit
tqdm
